The model library no longer depends on the IO library.
