+++ CMakeLists.txt
@@ -299,8 +299,8 @@ include_directories("${PROJECT_SOURCE_DIR}")
 add_subdirectory(core)
 add_subdirectory(geometry)
 add_subdirectory(solvers)
-add_subdirectory(io)
 add_subdirectory(model)
+add_subdirectory(io)
 
 # Now that we have gathered all our libraries, generate a polymec.cmake 
 # file that contains all the vital information.
