+++ io/silo_file.h
@@ -12,6 +12,7 @@
 #include "core/point_cloud.h"
 #include "core/slist.h"
 #include "geometry/polymesh.h"
+#include "model/neighbor_pairing.h"
 
 // Enables GZIP compression at the given level for Silo files. This is 
 // invoked globally and effects all file writes until it is set to a different 
@@ -281,5 +282,40 @@ exchanger_t* silo_file_read_exchanger(silo_file_t* file, const char* exchanger_n
 
 // Writes an exchanger object with the given name.
 void silo_file_write_exchanger(silo_file_t* file, const char* exchanger_name, exchanger_t* ex);
+
+// Returns true if the given silo_file contains a stencil with the given 
+// name, false otherwise.
+bool silo_file_contains_stencil(silo_file_t* file, const char* stencil_name);
+
+// Writes a stencil object with the given name.
+void silo_file_write_stencil(silo_file_t* file,
+                             const char* stencil_name,
+                             stencil_t* stencil);
+
+// This function extends the silo_file type to allow it to read in and 
+// return a newly-allocated stencil object from the entry in the 
+// file with the given name. The exchanger for the stencil is assigned
+// to the given MPI communicator.
+stencil_t* silo_file_read_stencil(silo_file_t* file,
+                                  const char* stencil_name,
+                                  MPI_Comm comm);
+
+// Returns true if the given silo file contains a neighbor_pairing object 
+// with the given name, false otherwise.
+bool silo_file_contains_neighbor_pairing(silo_file_t* file,
+                                         const char* neighbors_name);
+
+// Writes a neighbor_pairing object to an entry with the given name.
+void silo_file_write_neighbor_pairing(silo_file_t* file,
+                                      const char* neighbors_name,
+                                      neighbor_pairing_t* neighbors);
+
+// This function extends the silo_file type to allow it to read in and 
+// return a newly-allocated neighbor_pairing object from the entry in the 
+// file with the given name. The exchanger for the neighbor pairing is assigned
+// to the given MPI communicator.
+neighbor_pairing_t* silo_file_read_neighbor_pairing(silo_file_t* file,
+                                                    const char* neighbors_name,
+                                                    MPI_Comm comm);
 
 #endif
