+++ model/stencil.c
@@ -215,78 +215,12 @@ serializer_t* stencil_serializer()
   return serializer_new("stencil", stencil_byte_size, stencil_byte_read, stencil_byte_write, DTOR(stencil_free));
 }
 
-matrix_sparsity_t* sparsity_from_stencil(stencil_t* stencil)
+matrix_sparsity_t* matrix_sparsity_from_stencil(stencil_t* stencil)
 {
   adj_graph_t* g = stencil_as_graph(stencil);
   matrix_sparsity_t* sp = matrix_sparsity_from_graph(g, stencil->ex);
   adj_graph_free(g);
   return sp;
-}
-
-bool silo_file_contains_stencil(silo_file_t* file, const char* stencil_name)
-{
-  char name[FILENAME_MAX+1];
-  snprintf(name, FILENAME_MAX, "%s_stencil_name", stencil_name);
-  return silo_file_contains_string(file, name);
-}
-
-void silo_file_write_stencil(silo_file_t* file,
-                             const char* stencil_name,
-                             stencil_t* stencil)
-{
-  START_FUNCTION_TIMER();
-  char name_name[FILENAME_MAX+1];
-  snprintf(name_name, FILENAME_MAX, "%s_stencil_name", stencil_name);
-  silo_file_write_string(file, name_name, stencil->name);
-
-  char offsets_name[FILENAME_MAX+1];
-  snprintf(offsets_name, FILENAME_MAX, "%s_stencil_offsets", stencil_name);
-  silo_file_write_int_array(file, offsets_name, stencil->offsets, stencil->num_indices+1);
-
-  char indices_name[FILENAME_MAX+1];
-  snprintf(indices_name, FILENAME_MAX, "%s_stencil_indices", stencil_name);
-  silo_file_write_int_array(file, indices_name, stencil->indices, stencil->offsets[stencil->num_indices]);
-
-  if (stencil->ex != NULL)
-  {
-    char ex_name[FILENAME_MAX+1];
-    snprintf(ex_name, FILENAME_MAX, "%s_stencil_ex", stencil_name);
-    silo_file_write_exchanger(file, ex_name, stencil->ex);
-  }
-  STOP_FUNCTION_TIMER();
-}
-
-stencil_t* silo_file_read_stencil(silo_file_t* file,
-                                  const char* stencil_name,
-                                  MPI_Comm comm)
-{
-  START_FUNCTION_TIMER();
-  stencil_t* s = polymec_malloc(sizeof(stencil_t));
-  char name_name[FILENAME_MAX+1];
-  snprintf(name_name, FILENAME_MAX, "%s_stencil_name", stencil_name);
-  s->name = silo_file_read_string(file, name_name);
-
-  char offsets_name[FILENAME_MAX+1];
-  snprintf(offsets_name, FILENAME_MAX, "%s_stencil_offsets", stencil_name);
-  size_t size;
-  s->offsets = silo_file_read_int_array(file, offsets_name, &size);
-  s->num_indices = (int)(size) - 1;
-
-  if (s->offsets[s->num_indices] > 0)
-  {
-    char indices_name[FILENAME_MAX+1];
-    snprintf(indices_name, FILENAME_MAX, "%s_stencil_indices", stencil_name);
-    s->indices = silo_file_read_int_array(file, indices_name, &size);
-    ASSERT((int)size == s->offsets[s->num_indices]);
-  }
-  else
-    s->indices = NULL;
-
-  char ex_name[FILENAME_MAX+1];
-  snprintf(ex_name, FILENAME_MAX, "%s_stencil_ex", stencil_name);
-  s->ex = silo_file_read_exchanger(file, ex_name, comm);
-  STOP_FUNCTION_TIMER();
-  return s;
 }
 
 stencil_t* distance_based_point_stencil_new(point_cloud_t* points,
