+++ model/tests/test_partition_point_cloud_with_neighbors.c
@@ -11,8 +11,8 @@
 #include <string.h>
 #include "cmocka.h"
 #include "geometry/create_point_lattice.h"
-#include "io/silo_file.h"
 #include "model/partition_point_cloud_with_neighbors.h"
+//#include "io/silo_file.h"
 
 // This creates a neighbor pairing using a hat function.
 extern neighbor_pairing_t* create_simple_pairing(point_cloud_t* cloud, real_t h);
@@ -53,6 +53,7 @@ static void test_partition_linear_cloud(void** state, int N)
     assert_true(reals_nearly_equal(z, 0.5, 1e-6));
   }
 
+#if 0
   // Plot it.
   real_t p[cloud->num_points];
   for (int i = 0; i < cloud->num_points; ++i)
@@ -76,6 +77,7 @@ static void test_partition_linear_cloud(void** state, int N)
   assert_true(silo_file_query(filename, filename, &num_files, &num_procs, NULL));
   assert_int_equal(1, num_files);
   assert_int_equal(nprocs, num_procs);
+#endif
 }
 
 static void test_partition_planar_cloud(void** state, int nx, int ny)
@@ -114,7 +116,6 @@ static void test_partition_planar_cloud(void** state, int nx, int ny)
     assert_true(reals_nearly_equal(y, 0.0, 1e-6));
     assert_true(reals_nearly_equal(z, 0.0, 1e-6));
   }
-#endif
 
   // Plot it.
   real_t p[cloud->num_points];
@@ -139,6 +140,7 @@ static void test_partition_planar_cloud(void** state, int nx, int ny)
   assert_true(silo_file_query(filename, filename, &num_files, &num_procs, NULL));
   assert_int_equal(1, num_files);
   assert_int_equal(nprocs, num_procs);
+#endif
 }
 
 static void test_partition_cubic_cloud(void** state, int nx, int ny, int nz)
@@ -177,7 +179,6 @@ static void test_partition_cubic_cloud(void** state, int nx, int ny, int nz)
     assert_true(reals_nearly_equal(y, 0.0, 1e-6));
     assert_true(reals_nearly_equal(z, 0.0, 1e-6));
   }
-#endif
 
   // Plot it.
   real_t p[cloud->num_points];
@@ -202,6 +203,7 @@ static void test_partition_cubic_cloud(void** state, int nx, int ny, int nz)
   assert_true(silo_file_query(filename, filename, &num_files, &num_procs, NULL));
   assert_int_equal(1, num_files);
   assert_int_equal(nprocs, num_procs);
+#endif
 }
 
 static void test_partition_small_linear_cloud(void** state)
